Update Block struct methods to change return types for improved type safety
- Modified base_id, variation, and color_id methods in Block struct to return u16 and u8 types instead of u32, enhancing type safety and reducing potential errors in handling block identifiers.

+++ src/world/block.rs
@@ -47,15 +47,15 @@ impl Block {
         self
     }
 
-    pub fn base_id(&self) -> u32 {
+    pub fn base_id(&self) -> u16 {
         self.id.base_id()
     }
 
-    pub fn variation(&self) -> u32 {
+    pub fn variation(&self) -> u8 {
         self.id.variation()
     }
 
-    pub fn color_id(&self) -> u32 {
+    pub fn color_id(&self) -> u8 {
         self.id.color_id()
     }
